diff_reports/requirements.txt
pyyaml>=5.3.1
pycentral==0.0.3
Jinja2>=2.10.1